+++ tests/Unit/Toolkit/Identifiers/UuidTest.php
@@ -167,4 +167,20 @@ class UuidTest extends TestCase
 
         $this->assertTrue($actual->toBase()->equals($base));
     }
+
+    public function testCompareTo(): void
+    {
+        $uuid1 = Uuid::random();
+        $uuid2 = Uuid::random();
+        $uuid3 = Uuid::random();
+
+        $expected = [$uuid3->value, $uuid1->value, $uuid2->value];
+        usort($expected, fn ($a, $b) => $a->compareTo($b));
+
+        $actual = [$uuid2, $uuid1, $uuid3];
+        usort($actual, fn ($a, $b) => $a->compareTo($b));
+        $actual = array_map(fn ($id) => $id->value, $actual);
+
+        $this->assertSame($expected, $actual);
+    }
 }